requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
icalendar>=5.0.0
pytest>=7.4.0
pytest-django>=4.5.0
hypothesis>=6.82.0
django-cors-headers>=4.3.1
